Add options to specify bold/italic/bold-italic fonts
This patch adds options to specify different fonts for the bold, italic,
and bold-italic variants. These options can be useful, for example, if
you want to use a different bold style instead of the default. Since the
new options also allow you to disable bold and italic fonts by setting
them to normal, this patch removes the disable_bold and disable_italic
options.